requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
#python-slugify==4.0.1  # https://github.com/un33k/python-slugify
#Pillow==7.2.0  # https://github.com/python-pillow/Pillow
#argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
#whitenoise==5.2.0  # https://github.com/evansd/whitenoise
#redis==3.5.3  # https://github.com/andymccurdy/redis-py
#hiredis==1.1.0  # https://github.com/redis/hiredis-py
#celery==4.4.7  # pyup: < 5.0  # https://github.com/celery/celery
#django-celery-beat==2.0.0  # https://github.com/celery/django-celery-beat
#flower==0.9.5  # https://github.com/mher/flower
uvicorn==0.11.8  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.1.1  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
# django-allauth==0.42.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.2  # https://github.com/django-crispy-forms/django-crispy-forms
# django-redis==4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.11.1  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.4.0 # https://github.com/adamchainz/django-cors-headers
django-extensions==3.0.4  # https://github.com/django-extensions/django-extensions
ipython
jupyter
# django-typed-models==0.10.0
# django-polymorphic==3.0.0
# django-mptt==0.11.0
# django-polymorphic-tree==1.5.1
pandas==1.1.1